Address review: chunk sync_removed under SQLite's variable-param cap
A single `uuid IN (?, ?, …)` list sized to `removed_uuids.len()` could
exceed SQLITE_LIMIT_VARIABLE_NUMBER (default 999) on large diffs —
notably when `replace_books` clears an entire large library. Chunk at
500 (library_id + 1 per uuid → safely under the cap) and run both the
books_fts and books deletes per chunk inside the same transaction,
matching the existing tags/identifiers/backfill chunking pattern.
